@@ -1,130 +1,88 @@
 <?php
 
-namespace Sweetchuck\Robo\Serialize\Tests\Unit;
+declare(strict_types = 1);
 
-use Sweetchuck\Robo\Serialize\Task\SerializeTask;
-use Sweetchuck\Robo\Serialize\Test\Helper\Dummy\Subject01;
-use Sweetchuck\Robo\Serialize\Test\Helper\Dummy\Subject02;
+namespace Sweetchuck\Robo\Serialize\Tests\Unit\Task;
+
 use Codeception\Test\Unit;
-use Codeception\Util\Stub;
+use League\Container\Container as LeagueContainer;
+use League\Container\ContainerInterface;
+use Robo\Collection\CollectionBuilder;
+use Robo\Config\Config as RoboConfig;
 use Robo\Robo;
+use Sweetchuck\Codeception\Module\RoboTaskRunner\DummyOutput;
+use Sweetchuck\Robo\Serialize\Test\Helper\Dummy\DummyTaskBuilder;
+use Sweetchuck\Robo\Serialize\Test\UnitTester;
+use Symfony\Component\Console\Application as SymfonyApplication;
 use Symfony\Component\Console\Output\BufferedOutput;
-use Symfony\Component\Filesystem\Filesystem;
+use Symfony\Component\ErrorHandler\BufferingLogger;
 
 /**
  * @covers \Sweetchuck\Robo\Serialize\Task\SerializeTask
  */
 class SerializeTaskTest extends Unit
 {
-    /**
-     * @var \Sweetchuck\Robo\Serialize\Test\UnitTester
-     */
-    protected $tester;
+    protected UnitTester $tester;
 
-    /**
-     * {@inheritdoc}
-     */
-    protected function setUp()
+    protected ContainerInterface $container;
+
+    protected RoboConfig $config;
+
+    protected CollectionBuilder $builder;
+
+    protected DummyTaskBuilder $taskBuilder;
+
+    protected function _before()
     {
-        $tmpDir = codecept_output_dir('tmp');
-        if (!is_dir($tmpDir)) {
-            mkdir($tmpDir, 0777 - umask(), true);
-        }
+        parent::_before();
 
-        parent::setUp();
-    }
+        Robo::unsetContainer();
 
-    // @codingStandardsIgnoreStart
-    /**
-     * {@inheritdoc}
-     */
-    public function _after()
-    {
-        // @codingStandardsIgnoreEnd
-        parent::_after();
+        $this->container = new LeagueContainer();
+        $application = new SymfonyApplication('Sweetchuck - Robo Git', '1.0.0');
+        $this->config = new RoboConfig();
+        $input = null;
+        $output = new DummyOutput([
+            'verbosity' => DummyOutput::VERBOSITY_DEBUG,
+        ]);
 
-        $tmpDir = codecept_output_dir('tmp');
-        if (is_dir($tmpDir)) {
-            (new Filesystem())->remove($tmpDir);
-        }
-    }
+        $this->container->add('container', $this->container);
 
-    public function testGetSetSerializer(): void
-    {
-        $task = new SerializeTask();
+        Robo::configureContainer($this->container, $application, $this->config, $input, $output);
+        $this->container->share('logger', BufferingLogger::class);
 
-        $this->tester->assertEquals('json', $task->getSerializer(), 'Default value: json');
-
-        $task->setSerializer('yaml');
-        $this->tester->assertEquals('yaml', $task->getSerializer(), 'New value: yaml');
-
-        $task->setSerializer('yml');
-        $this->tester->assertEquals('yml', $task->getSerializer(), 'New value: yml');
-
-        try {
-            $task->setSerializer('none');
-            $this->tester->fail("Invalid serializer name wasn't detected");
-        } catch (\InvalidArgumentException $e) {
-            $this->tester->assertTrue(true, 'Invalid serializer name was detected');
-        }
-    }
-
-    public function testGetSetDestination(): void
-    {
-        $task = new SerializeTask();
-
-        $this->tester->assertEquals(STDOUT, $task->getDestination(), 'Default value: STDOUT');
-
-        $task->setDestination('foo.yml');
-        $this->tester->assertEquals('foo.yml', $task->getDestination(), 'New value: foo.yml');
-
-        $task->setDestination(STDERR);
-        $this->tester->assertEquals(STDERR, $task->getDestination(), 'New value: STDERR');
-
-        $bufferedOutput = new BufferedOutput();
-        $task->setDestination($bufferedOutput);
-        $this->tester->assertEquals($bufferedOutput, $task->getDestination(), 'New value: BufferedOutput');
-
-        try {
-            $task->setDestination([]);
-            $this->tester->fail("Invalid destination wasn't detected");
-        } catch (\InvalidArgumentException $e) {
-            $this->tester->assertTrue(true, 'Invalid destination was detected');
-        }
-    }
-
-    public function testOptions(): void
-    {
-        $options = [
-            'subject' => ['a' => 'b'],
-            'serializer' => 'yaml',
-            'destination' => 'foo.yml',
-        ];
-
-        $task = new SerializeTask($options);
-        $this->tester->assertEquals($options['subject'], $task->getSubject());
-        $this->tester->assertEquals($options['serializer'], $task->getSerializer());
-        $this->tester->assertEquals($options['destination'], $task->getDestination());
+        $this->builder = CollectionBuilder::create($this->container, null);
+        $this->taskBuilder = new DummyTaskBuilder();
+        $this->taskBuilder->setContainer($this->container);
+        $this->taskBuilder->setBuilder($this->builder);
     }
 
     public function casesRun(): array
     {
-        $subjectSimple = ['foo' => 'bar'];
-        $subjectDeep = new \stdClass();
-        $subjectDeep->foo = new \stdClass();
-        $subjectDeep->foo->bar = new \stdClass();
-        $subjectDeep->foo->bar->baz = new \stdClass();
-        $subjectDeep->foo->bar->baz->a = 'b';
-        $subjectObject = [
-            'normal' => new Subject01(),
-            'serializable' => new Subject02(['d' => 'e']),
-            'r1' => STDERR,
-        ];
-
-        $null = function_exists('yaml_emit') ? '~' : 'null';
-
         return [
-            'simple - json;' => [
+            'basic pecl_yaml' => [
+                implode("\n", [
+                    '---',
+                    'foo: bar',
+                    '...',
+                    '',
+                ]),
+                'pecl_yaml',
+                [
+                    'foo' => 'bar',
+                ],
+            ],
+            'basic symfony_yaml' => [
+                implode("\n", [
+                    'foo: bar',
+                    '',
+                ]),
+                'symfony_yaml',
+                [
+                    'foo' => 'bar',
+                ],
+            ],
+            'basic json' => [
                 implode("\n", [
                     '{',
                     '    "foo": "bar"',
@@ -132,76 +90,9 @@ class SerializeTaskTest extends Unit
                     '',
                 ]),
                 'json',
-                $subjectSimple,
-            ],
-            'simple - yaml' => [
-                implode("\n", [
-                    '---',
-                    'foo: bar',
-                    '...',
-                    '',
-                ]),
-                'yaml',
-                $subjectSimple,
-            ],
-            'deep - json;' => [
-                implode("\n", [
-                    '{',
-                    '    "foo": {',
-                    '        "bar": {',
-                    '            "baz": {',
-                    '                "a": "b"',
-                    '            }',
-                    '        }',
-                    '    }',
-                    '}',
-                    '',
-                ]),
-                'json',
-                $subjectDeep,
-            ],
-            'deep - yaml' => [
-                implode("\n", [
-                    '---',
-                    'foo:',
-                    '  bar:',
-                    '    baz:',
-                    '      a: b',
-                    '...',
-                    '',
-                ]),
-                'yaml',
-                $subjectDeep,
-            ],
-            'object - json' => [
-                implode("\n", [
-                    '{',
-                    '    "normal": {',
-                    '        "myPublic": "a"',
-                    '    },',
-                    '    "serializable": {',
-                    '        "d": "e"',
-                    '    },',
-                    '    "r1": null',
-                    '}',
-                    '',
-                ]),
-                'json',
-                $subjectObject,
-            ],
-            'object - yaml' => [
-                implode("\n", [
-                    '---',
-                    'normal:',
-                    '  myPublic: a',
-                    'serializable:',
-                    '  d: e',
-                    "r1: $null",
-                    '...',
-                    '',
-                ]),
-                'yaml',
-                $subjectObject,
+                [
+                    'foo' => 'bar',
+                ],
             ],
         ];
     }
@@ -209,72 +100,18 @@ class SerializeTaskTest extends Unit
     /**
      * @dataProvider casesRun
      */
-    public function testRunOutput(string $expected, string $serializer, $subject): void
+    public function testRun(string $expected, string $serializerName, $value): void
     {
-        $destination = new BufferedOutput();
+        $writer = new BufferedOutput();
 
-        $task = $this->getTask();
-        $result = $task
-            ->setSubject($subject)
-            ->setSerializer($serializer)
-            ->setDestination($destination)
-            ->run();
+        $task = $this->taskBuilder->taskSerialize([
+            'value' => $value,
+            'serializer' => $this->taskBuilder->getSerializer($serializerName),
+            'writer' => $writer,
+        ]);
 
-        $this->tester->assertEquals(0, $result->getExitCode());
-        $this->tester->assertEquals($expected, $destination->fetch());
-    }
+        $result = $task->run();
 
-    /**
-     * @dataProvider casesRun
-     */
-    public function testRunString(string $expected, string $serializer, $subject): void
-    {
-        $destination = codecept_output_dir('tmp/destination.txt');
-        $this->tester->assertFileNotExists($destination);
-
-        $task = $this->getTask();
-        $result = $task
-            ->setSubject($subject)
-            ->setSerializer($serializer)
-            ->setDestination($destination)
-            ->run();
-
-        $this->tester->assertEquals(0, $result->getExitCode());
-        $this->tester->assertFileExists($destination);
-        $this->tester->assertEquals($expected, file_get_contents($destination));
-    }
-
-    /**
-     * @dataProvider casesRun
-     */
-    public function testRunResource(string $expected, string $serializer, $subject): void
-    {
-        $fileName = codecept_output_dir('tmp/destination.txt');
-        $this->tester->assertFileNotExists($fileName);
-        $destination = fopen($fileName, 'w');
-
-        $task = $this->getTask();
-        $result = $task
-            ->setSubject($subject)
-            ->setSerializer($serializer)
-            ->setDestination($destination)
-            ->run();
-
-        $this->tester->assertEquals(0, $result->getExitCode());
-        $this->tester->assertFileExists($fileName);
-        $this->tester->assertEquals($expected, file_get_contents($fileName));
-    }
-
-    protected function getTask(): SerializeTask
-    {
-        $container = Robo::createDefaultContainer();
-        Robo::setContainer($container);
-
-        /** @var \Sweetchuck\Robo\Serialize\Task\SerializeTask $task */
-        $task = Stub::construct(SerializeTask::class, [[], []]);
-
-        $task->setLogger($container->get('logger'));
-
-        return $task;
+        $this->tester->assertEquals($expected, $writer->fetch());
     }
 }
