@@ -1,26 +1,14 @@
 <?php
 
-namespace Sweetchuck\Robo\Serialize\Tests\Acceptance;
+declare(strict_types = 1);
+
+namespace Sweetchuck\Robo\Serialize\Tests\Acceptance\Task;
 
 use Sweetchuck\Robo\Serialize\Test\AcceptanceTester;
-use Sweetchuck\Robo\Serialize\Test\Helper\RoboFiles\SerializeTaskRoboFile;
+use Sweetchuck\Robo\Serialize\Test\Helper\RoboFiles\RoboFileSerialize;
 
 class SerializeTaskCest
 {
-    public function testSerializeYaml(AcceptanceTester $I): void
-    {
-        $id = 'serialize:yaml';
-        $expectedStdOutput = implode("\n", [
-            '---',
-            'a: b',
-            '...',
-            '',
-        ]);
-        $I->runRoboTask($id, SerializeTaskRoboFile::class, 'serialize:yaml');
-        $I->assertEquals(0, $I->getRoboTaskExitCode($id));
-        $I->assertEquals($expectedStdOutput, $I->getRoboTaskStdOutput($id));
-    }
-
     public function testSerializeJson(AcceptanceTester $I): void
     {
         $id = 'serialize:json';
@@ -30,7 +18,7 @@ class SerializeTaskCest
             '}',
             '',
         ]);
-        $I->runRoboTask($id, SerializeTaskRoboFile::class, 'serialize:json');
+        $I->runRoboTask($id, RoboFileSerialize::class, 'serialize', 'json');
         $I->assertEquals(0, $I->getRoboTaskExitCode($id));
         $I->assertEquals($expectedStdOutput, $I->getRoboTaskStdOutput($id));
     }
